% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{counterparts}
\alias{counterparts}
\title{Internal Counterparts Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 303 rows and 5 columns.
}
\usage{
counterparts
}
\description{
This dataset is accessible through \link{ids_list_counterparts}.
}
\keyword{internal}
