% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{entities}
\alias{entities}
\title{Internal Entities Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 134 rows and 17 columns.
}
\usage{
entities
}
\description{
This dataset is accessible through \link{ids_list_entities}.
}
\keyword{internal}
