% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{geographies}
\alias{geographies}
\title{Internal Geographies Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 136 rows and 14 columns.
}
\usage{
geographies
}
\description{
This dataset is accessible through \link{ids_list_geographies}.
}
\keyword{internal}
