% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bulk_info.R
\name{read_bulk_info}
\alias{read_bulk_info}
\title{Read Bulk Information from World Bank Data Catalog API}
\usage{
read_bulk_info()
}
\value{
A list containing metadata about available bulk downloads
}
\description{
Internal function to retrieve bulk download information for International
Debt Statistics (IDS) from the World Bank Data Catalog API.
}
\keyword{internal}
