% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{series}
\alias{series}
\title{Internal Series Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 574 rows and 6 columns.
}
\usage{
series
}
\description{
This dataset is accessible through \link{ids_list_series}.
}
\keyword{internal}
