% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{series_topics}
\alias{series_topics}
\title{Internal Series-Topics Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 598 rows and 3 columns.
}
\usage{
series_topics
}
\description{
This dataset is accessible through \link{ids_list_series_topics}.
}
\keyword{internal}
