% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbids-package.R
\docType{package}
\name{wbids-package}
\alias{wbids}
\alias{wbids-package}
\title{wbids: Seamless Access to World Bank International Debt Statistics (IDS)}
\description{
Access and analyze the World Bank's International Debt Statistics (IDS) \url{https://www.worldbank.org/en/programs/debt-statistics/ids}. IDS provides creditor-debtor relationships between countries, regions, and institutions. 'wbids' enables users to download, process and work with IDS series across multiple geographies, counterparts, and time periods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://teal-insights.github.io/r-wbids/}
  \item \url{https://github.com/teal-insights/r-wbids/}
  \item Report bugs at \url{https://github.com/teal-insights/r-wbids/issues}
}

}
\author{
\strong{Maintainer}: Teal Emery \email{lte@tealinsights.com}

Authors:
\itemize{
  \item Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID})
}

Other contributors:
\itemize{
  \item Teal Insights [copyright holder]
  \item Christopher Smith \email{chriscarrollsmith@gmail.com} [contributor]
}

}
\keyword{internal}
